% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/brackets.R
\name{brace}
\alias{brace}
\title{Add a Curly Brace to a Plot}
\usage{
brace(x1, y1, x2, y2, h = 1, prop = FALSE, code = 1, aspect = 0.5,
  length = 0.125, curvature = 0.5, ...)
}
\arguments{
\item{x1}{x coordinate of point *from* which to draw.}

\item{y1}{y coordinate of point *from* which to draw.}

\item{x2}{x coordinate of point *to* which to draw.}

\item{y2}{y coordinate of point *to* which to draw.}

\item{h}{height}

\item{prop}{logical; if `TRUE', the height is propotional to the width of
the brace; if `FALSE', the height is constant (picas).}

\item{code}{integer code. determine direction of the curly brace}

\item{aspect}{determines the position of the break.  produces balanced curly brace with the default aspect (0.5).}

\item{length}{length of curls [0, 1]}

\item{curvature}{roundness [0, 1]}

\item{...}{other graphical parameters}
}
\description{
Draw curly braces between a pair of points
}
\examples{
plot(0, type="n", xlim=c(0,20), ylim=c(0,20))
for(i in 1:5) {
   brace(i, 10-i, i, 10+i, col=i, h=1)
   brace(10-i, 20-i, 10+i, 20-i)
   brace(6, 14-2*i, 6+2*i, 14)
}
}

