% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/brackets.R
\name{grid.bracket}
\alias{bracketGrob}
\alias{grid.bracket}
\title{Draw a Brackets}
\usage{
grid.bracket(x1 = 0, y1 = 0, x2 = 1, y2 = 1, center = 0.5,
  type = c("curly", "round", "square", "angle"), head = list(height = c(0.5,
  0.5), width = c(1/8, 1/8), curvature = c(0.5, 0.5)), tail = list(height =
  c(0.5, 0.5), width = c(1/8, 1/8), curvature = c(0.5, 0.5)),
  direction = c("above", "under", "left", "right"), default.units = "npc",
  gp = gpar())

bracketGrob(x1 = 0, y1 = 0, x2 = 1, y2 = 1, center = 0.5,
  type = c("curly", "round", "square", "angle"), head = list(height = c(0.5,
  0.5), width = c(1/8, 1/8), curvature = c(0.5, 0.5)), tail = list(height =
  c(0.5, 0.5), width = c(1/8, 1/8), curvature = c(0.5, 0.5)),
  direction = c("above", "under", "left", "right"), default.units = "npc",
  gp = gpar())
}
\arguments{
\item{x1}{x coordinate of bottom left}

\item{y1}{y coordinate of bottom left}

\item{x2}{x coordinate of top right}

\item{y2}{y coordinate of top right}

\item{center}{relative center location, where head is located}

\item{type}{bracket type: "curly", "round", "square", or "angle".}

\item{head}{shape of the head. height of the head. widths and curvatures of
left and right sides. if NULL, no head.}

\item{tail}{shape of the left and right tails.}

\item{direction}{direction of the bracket}

\item{default.units}{default units}

\item{gp}{An object of class `gpar'. graphical parameters.}
}
\description{
Draw brackets, parentheses, curly brackets, angled brackets
}
\examples{
grid.newpage()
grid.bracket()
}

