% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGstats.R
\name{CGstats_internal}
\alias{CGstats_internal}
\title{res  <- c(ans, list(cont.names=cont.names, disc.names=disc.names, disc.levels=disc.levels))
##class(res) <- "CGstats"
return(res)}
\usage{
CGstats_internal(object, disc.names = NULL, cont.names = NULL,
  homogeneous = TRUE, simplify = TRUE)
}
\description{
res  <- c(ans, list(cont.names=cont.names, disc.names=disc.names, disc.levels=disc.levels))
##class(res) <- "CGstats"
return(res)
}

