% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citest_2025.r, R/citest_generic.R
\name{citest-generic}
\alias{citest-generic}
\alias{ci_test}
\alias{ciTest}
\title{Test for conditional independence in contingecy table.}
\usage{
ci_test(x, set = NULL, ...)

ciTest(x, set = NULL, ...)
}
\arguments{
\item{x}{An object for which a test for conditional independence is to be
made. See 'details' for valid types of \code{x}.}

\item{set}{A specification of the test to be made. The tests are of the form
u and v are independent condionally on S where u and v are variables and
S is a set of variables. See 'details' for details about specification of
\code{set}.}

\item{\dots}{Additional arguments to be passed on to other methods.}
}
\value{
An object of class \code{citest} (which is a list).

An object of class \code{citest} (which is a list).
}
\description{
Generic function for conditional independence test. Specializes
to specific types of data.
}
\details{
\code{set} can be
\enumerate{
\item a vector,
\item a right-hand sided
formula in which variables are separated by '+'.
}

In either case, it is tested if the first two variables in the
\code{set} are conditionally independent given the remaining
variables in \code{set}.  (Notice an abuse of the '+' operator in
the right-hand sided formula: The order of the variables does
matter.)

\code{x} can be
\enumerate{
\item a table (an array). In this case \code{ciTest_table} is called.
\item a dataframe whose columns are numerics and factors. In this
case \code{ciTest_df} is called.
\item a list with components \code{cov} and \code{n.obs}. In this
case \code{ciTest_mvn} is called.
}

\code{set} can be
\enumerate{
\item a vector,
\item a right-hand sided
formula in which variables are separated by '+'.
}

In either case, it is tested if the first two variables in the
\code{set} are conditionally independent given the remaining
variables in \code{set}.  (Notice an abuse of the '+' operator in
the right-hand sided formula: The order of the variables does
matter.)
}
\examples{

data(lizard)

## lizard is has named dimnames
names(dimnames(lizard))
## checked with
is.named.array(lizard)

## Testing for conditional independence:
# the following are all equivalent:
ciTest(lizard, set=~diam + height + species)
# ciTest(lizard, set=c("diam", "height", "species"))
# ciTest(lizard, set=1:3)
# ciTest(lizard)
# (The latter because the names in lizard are as given above.)

## Testing for marginal independence
ciTest(lizard, set= ~ diam + height)
ciTest(lizard, set=1:2)

## Getting slice information:
ciTest(lizard, set=c("diam", "height", "species"), slice.info=TRUE)$slice

## Do Monte Carlo test instead of usual likelihood ratio test. 

ciTest(lizard, set=c("diam", "height", "species"), method="mc", B=400)

## contingency table:
data(reinis)
## dataframe with only numeric variables:
data(carcass)
## dataframe with numeric variables and factors:
data(milkcomp1)

ciTest(cov.wt(carcass, method='ML'), set=~Fat11 + Meat11 + Fat12)
ciTest(reinis, set=~smo + phy + sys)
ciTest(milkcomp1, set=~tre + fat + pro)
}
\seealso{
\code{\link{chisq.test}}

\code{\link{ciTest_table}},
\code{\link{ciTest_df}},
\code{\link{ciTest_mvn}},
\code{\link{chisq.test}}
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{htest}
