% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciTest.R
\name{ciTest_table}
\alias{ciTest_table}
\title{Test for conditional independence in a contingency table}
\usage{
ciTest_table(x, set = NULL, statistic = "dev", method = "chisq",
  adjust.df = TRUE, slice.info = TRUE, L = 20, B = 200, ...)
}
\arguments{
\item{x}{A contingency table.}

\item{...}{Additional arguments.}
}
\value{
An object of class 'citest' (which is a list).
}
\description{
Test for conditional independence in a contingency table
}
\details{
\code{set} can be 1) a vector or 2) a right-hand sided formula in which
variables are separated by '+'. In either case, it is tested if the first
two variables in the \code{set} are conditionally independent given the
remaining variables in \code{set}.  (Notice an abuse of the '+' operator in
the right-hand sided formula: The order of the variables does matter.)

If \code{set} is \code{NULL} then it is tested whether the first two
variables are conditionally independent given the remaining variables.
}
\examples{

data(reinis)
ciTest(reinis, set=~smo+phy+sys)
ciTest_table(reinis, set=~smo+phy+sys)

}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\seealso{
\code{\link{ciTest}}, \code{\link{ciTest.data.frame}},
    \code{\link{ciTest_df}}, \code{\link{ciTest.list}},
    \code{\link{ciTest_mvn}}, \code{\link{chisq.test}}
}
\keyword{htest}

