\name{dModel-class}
\Rdversion{1.1}
\docType{class}
\alias{dModel-class}
\alias{cModel-class}
\alias{mModel-class}

\title{Class \code{"dModel"}}
\description{
  Setting formal classes for dModel, cModel and mModel objects
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{cModel}"}, directly.
Class \code{"\linkS4class{mModel}"}, by class "cModel", distance 2.
Class \code{"\linkS4class{oldClass}"}, by class "cModel", distance 3.
}
\section{Methods}{
No methods defined with class "dModel" in the signature.
}
%\references{
%%  ~~put references to the literature/web site here~~
%}
\author{
  Søren Højsgaard, sorenh at mail dot dk
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
\examples{
showClass("dModel")
}
\keyword{classes}
