% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_grips.R
\name{emat_operations}
\alias{emat_operations}
\alias{emat_compare}
\alias{emat_complement}
\alias{emat_sort}
\alias{order_rows}
\title{Edge matrix operations}
\usage{
emat_compare(emat1, emat2)

emat_complement(emat1, emat2)

emat_sort(emat1)

order_rows(emat)
}
\arguments{
\item{emat, emat1, emat2}{Edge matrix (a 2 x p matrix)}
}
\description{
Edge matrix operations needed for ips algorithms
}
\details{
An emat with p edges is represented by a 2 x p matrix.
}
\note{
These functions may well be removed from the package in future relases
}
\examples{
emat1 <- model_saturated(3:4, type="emat")
emat2 <- model_saturated(1:4, type="emat")
emat_complement(emat1, emat2)
emat3 <- model_saturated(2:4, type="emat")
emat_compare(emat1, emat3)
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
