% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_grips.R
\name{fast_cov}
\alias{fast_cov}
\title{Fast computation of covariance / correlation matrix}
\usage{
fast_cov(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{a numeric matrix(like object).}

\item{center, scale}{Should columns in x be centered and/or scaled}
}
\description{
Fast computation of covariance / correlation matrix
}
