% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_grips.R
\name{generate_models}
\alias{generate_models}
\alias{emat_saturated_model}
\alias{model_saturated}
\alias{model_random_tree}
\alias{model_rectangular_grid}
\alias{model_line}
\alias{model_star}
\alias{model_loop}
\alias{model_random}
\title{Generate various grapical models}
\usage{
emat_saturated_model(index)

model_saturated(index, type = "emat", nms = NULL)

model_random_tree(index, prob = 0, type = "emat", nms = NULL)

model_rectangular_grid(dim, type = "emat", nms = NULL)

model_line(index, type = "emat", nms = NULL)

model_star(index, type = "emat", nms = NULL)

model_loop(index, prob = 0, type = "emat", nms = NULL)

model_random(index, prob = 0.1, type = "emat", nms = NULL)
}
\arguments{
\item{index}{A vector of integers}

\item{type}{Output type.}

\item{nms}{Names of variables.}

\item{prob}{Probability of any edge being present.}

\item{dim}{A vector with dimensions}
}
\description{
Models are represented in various forms
}
