% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_grips.R
\name{generate_n01}
\alias{generate_n01}
\title{Genrate matrix of N(0, 1) variables}
\usage{
generate_n01(n.obs, nvar, seed = 2022)
}
\arguments{
\item{n.obs}{Number of rows}

\item{nvar}{Number of columns}

\item{seed}{Seed for random number generator}
}
\description{
Genrate matrix of N(0, 1) variables
}
