% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI-xxx2yyyParms.R
\name{ghk2phkParms}
\alias{ghk2phkParms}
\alias{ghk2pmsParms}
\alias{phk2ghkParms}
\alias{phk2pmsParms}
\alias{pms2ghkParms}
\alias{pms2phkParms}
\title{Conversion between different parametrizations of mixed interaction models.}
\usage{
ghk2phkParms(parms)
}
\arguments{
\item{parms}{Parameters of a mixed interaction model}
}
\value{
Parameters of a mixed interaction model.  %% ~Describe the value
returned %% If it is a LIST, use %% \item{comp1 }{Description of 'comp1'} %%
\item{comp2 }{Description of 'comp2'} %% ...
}
\description{
Functions to convert between canonical parametrization (g,h,K), moment
parametrization (p,m,S) and mixed parametrization (p,h,K).
}
\author{
S<f8>ren H<f8>jsgaard, \email{sorenh@math.aau.dk}
}
\keyword{utilities}

