\name{gRim-internal}

\alias{dModel}
\alias{cModel}
\alias{mModel}



%% iModel methods (some needs .Rd-files)
\alias{logLik.iModel}
\alias{summary.iModel}
\alias{print.iModelsummary}
\alias{update.iModel}
\alias{extractAIC.iModel}

%% Printing
\alias{print.iModel}

%% plot methods
\alias{plot.mModel}
\alias{plot.iModel}
\alias{iplot.iModel}
\alias{iplot.mModel}

%% model fitting
\alias{fit.cModel}
\alias{fit.dModel}
\alias{fit.mModel}
\alias{fitted.dModel}

%% FIXME
\alias{logLik.mModel} %% Bør erstattes med iModel metoden
\alias{update.mModel} %% Bør erstattes med iModel metoden
\alias{formula.iModel}
\alias{terms.iModel}

%% Needs .Rd files
\alias{CItest.data.frame}
\alias{compareModels.iModel}
\alias{print.compareiModels}
\alias{CGstats2mmodParms}
\alias{moment2pmsParms}

%% TO BE SORTED
\alias{compareGC}
\alias{compareGCpairs}
\alias{pFormula}
\alias{print.MIfit}
\alias{print.MIparms}
\alias{weakMarginalModel}
\alias{weakMarginalData}

\title{Internal functions in the gRim package}
\description{
  Internal functions called by other functions.
}
\keyword{internal}
