% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/brackets.R
\name{grim.bracket}
\alias{grim.bracket}
\title{Add a Bracket to a Plot (base graphics)}
\usage{
grim.bracket(x1, y1, x2, y2, h = unit(1, "picas"), prop = FALSE, code = 1,
  offset = 0, padding = rep(0, 4), center = 0.5, type = c("curly",
  "round", "square", "angle"), head = list(height = c(0.5, 0.5), width =
  c(1/8, 1/8), curvature = c(0.5, 0.5)), tail = list(height = c(0.5, 0.5),
  width = c(1/8, 1/8), curvature = c(0.5, 0.5)), ...)
}
\arguments{
\item{x1}{x coordinate of point *from* which to draw.}

\item{y1}{y coordinate of point *from* which to draw.}

\item{x2}{x coordinate of point *to* which to draw.}

\item{y2}{y coordinate of point *to* which to draw.}

\item{h}{height}

\item{prop}{logical; if `TRUE', the height is propotional to the width of
the bracket; if `FALSE', the height is constant (picas).}

\item{code}{integer code. determine direction of the bracket left/above (1), right/under (2)}

\item{offset}{offset from the baseline (x1, y1) -- (x2, y2)}

\item{padding}{padding sizes of bottom, left, top, right}

\item{center}{relative center location, where head is located}

\item{type}{bracket type: "curly", "round", "square", or "angle".}

\item{head}{shape of the head. height of the head. widths and curvatures of
left and right sides. if NULL, no head.}

\item{tail}{shape of the left and right tails.}

\item{...}{other graphical parameters}
}
\description{
Draw brackets, parentheses, curly brackets, angled brackets between a pair of points
}
\examples{
plot(0, type="n", xlim=c(-1, 16), ylim=c(-0.5, 3.5), xaxp=c(0,15,15), yaxp=c(0,3,3), ylab="")
grim.bracket(0, 0, 0, 3)
grim.bracket(1, 0, 1, 3, h=2)
grim.bracket(2, 0, 2, 3, h=3)
grim.bracket(3, 0, 3, 3, code=2)
grim.bracket(4, 0, 4, 3, center=0.3)
grim.bracket(5, 0, 5, 3, center=0.9)
grim.bracket(6, 0, 6, 3, lty=2)
grim.bracket(7, 0, 7, 3, lwd=2)
grim.bracket(8, 0, 8, 3, col="red")
grim.bracket(9, 0, 9, 3, type="round")
grim.bracket(10, 0, 10, 3, type="round", h=3)
grim.bracket(11, 0, 11, 3, type="square", h=2)
grim.bracket(12, 0, 12, 3, type="square", code=2)
grim.bracket(13, 0, 13, 3, type="angle")
grim.bracket(14, 0, 14, 3, type="angle", code=2, center=1/3, h=2)
grim.bracket(15, 0, 15, 3, type="angle", code=2, center=2/3)
}

