% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qaudrilaterals.R
\name{grim.rect}
\alias{grim.rect}
\title{Add Rectangles to a Base Graphic Plot}
\usage{
grim.rect(x0, y0, x1, y1, hpos = 0, hoffset = 0, voffset = 0, width = 1,
  height = 1, just = "center", hjust = NULL, vjust = NULL, ...)
}
\arguments{
\item{x0}{x coordinate of the origin}

\item{y0}{y coordinate of the origin}

\item{x1}{x coordinate of the unit vector}

\item{y1}{y coordinate of the unit vector}

\item{hpos}{horizontal position of rectangle on the baseline (x0, y0) -- (x1, y1)}

\item{hoffset}{horizontal offset (in picas if unit is not specified) from the \code{hpos}}

\item{voffset}{vertical offset (in picas if unit is not specified) from the baseline (x0, y0) -- (x1, y1)}

\item{width}{width of the rectangle (relative to the length of the unit vector)}

\item{height}{height of the rectangle (in picas)}

\item{just}{the justification. See \code{\link[grid]{grid.rect}}}

\item{hjust}{A numeric vector specifying horizontal justification. See \code{\link[grid]{grid.rect}}}

\item{vjust}{A numeric vector specifying vertical justification. See \code{\link[grid]{grid.rect}}}

\item{...}{other graphical parameters}
}
\description{
Draw rectangles
}

