% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text.R
\name{grim.text}
\alias{grim.text}
\title{Add Text to a Base Graphics Plot}
\usage{
grim.text(x0, y0, x1, y1, label, hpos = 0.5, hoffset = 0, voffset = 0,
  just = "center", hjust = NULL, vjust = NULL, rot = 0, ...)
}
\arguments{
\item{x0}{x coordinate of the origin}

\item{y0}{y coordinate of the origin}

\item{x1}{x coordinate of the unit vector}

\item{y1}{y coordinate of the unit vector}

\item{label}{text labels}

\item{hpos}{horizontal position of text on the baseline (x0, y0) -- (x1, y1)}

\item{hoffset}{horizontal offset (in picas if unit is not specified) from the
\code{hpos}}

\item{voffset}{vertical offset (in picas if unit is not specified) from the
baseline (x0, y0) -- (x1, y1)}

\item{just}{the justification of the text. If there are two values, the
first value specifies horizontal justification and the second value
specifies vertical justification. Possible string values are: vertical
justification. Possible string values are: \code{left}, \code{right},
\code{centre}, \code{center}, \code{bottom}, and \code{top}. See \link[grid]{grid.text}}

\item{hjust}{A numeric vector specifying horizontal justification. See \link[grid]{grid.text}}

\item{vjust}{A numeric vector specifying vertical justification. See \link[grid]{grid.text}}

\item{rot}{the angle to rotatate the text. See \link[grid]{grid.text}}

\item{...}{other graphical parameters}
}
\description{
Draw rotated text
}
\examples{
xs <- 0:10
ys <- 10*runif(11)
labels <- signif(sqrt(diff(xs)^2 + diff(ys)^2), 3)
plot(xs, ys, xlim = c(0,10), ylim = c(0,10))
for(i in 1:10) {
   grim.bracket(xs[i], ys[i], xs[i+1], ys[i+1],voffset = 0.2, code = i\%\%2 + 1)
   grim.text(xs[i], ys[i], xs[i+1], ys[i+1], label = labels[i],voffset = 2*ifelse(i\%\%2 == 0, 1, -1))
}
}

