% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iModel-general.R
\name{iModel-general}
\alias{extractAIC.iModel}
\alias{formula.iModel}
\alias{iModel-general}
\alias{isDecomposable.dModel}
\alias{isGraphical.dModel}
\alias{logLik.iModel}
\alias{modelProperties}
\alias{modelProperties.dModel}
\alias{print.iModelsummary}
\alias{summary.iModel}
\alias{terms.iModel}
\title{General functions related to iModels}
\usage{
\method{logLik}{iModel}(object, ...)

\method{extractAIC}{iModel}(fit, scale, k = 2, ...)

\method{summary}{iModel}(object, ...)

\method{print}{iModelsummary}(x, ...)

\method{formula}{iModel}(x, ...)

\method{terms}{iModel}(x, ...)

\method{isGraphical}{dModel}(x)

\method{isDecomposable}{dModel}(x)

modelProperties(object)

\method{modelProperties}{dModel}(object)
}
\arguments{
\item{object, fit, x}{An \code{iModel} object.}

\item{...}{Currently unused.}

\item{scale}{Unused (and irrelevant for these models)}

\item{k}{Weight of the degrees of freedom in the AIC formula}
}
\description{
General functions related to iModels
}

