% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_grim.R
\name{imodel-info}
\alias{imodel-info}
\alias{getmi}
\title{Get information about mixed interaction model objects}
\usage{
getmi(object, name)
}
\arguments{
\item{object}{An \code{iModel} object.}

\item{name}{The slot / information to be extracted.}
}
\description{
General functions related to iModels
}
