% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imodel-mmod.R
\name{imodel-mmod}
\alias{imodel-mmod}
\alias{mmod}
\alias{coef.mModel}
\alias{coefficients.mModel}
\alias{print.mModel}
\alias{summary.mModel}
\alias{mmod_dimension}
\title{Mixed interaction model.}
\usage{
mmod(formula, data, marginal = NULL, fit = TRUE, details = 0)
}
\arguments{
\item{formula}{A right hand sided formula specifying the model.}

\item{data}{Data (a dataframe)}

\item{marginal}{A possible subsets of columns of \code{data}; useful when
\code{formula} contains model specification shortcuts.}

\item{fit}{Currently not used}

\item{details}{For printing debugging information}
}
\value{
An object of class \code{mModel} and the more general class
\code{iModel}.
}
\description{
A mixed interaction model is a model (often with conditional
independence restrictions) for a combination of discrete and continuous
variables.
}
\examples{

### FIXME: To be written
}
\seealso{
\code{\link{dmod}}, \code{\link{cmod}}.
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{models}
