% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_grips.R
\name{impose_zero}
\alias{impose_zero}
\title{Impose zeros in matrix entries which do not correspond to an edge.}
\usage{
impose_zero(emat, K)
}
\arguments{
\item{emat}{Edge matrix (2 x p matrix)}

\item{K}{Matrix; typically a concentration matrix.}
}
\description{
Impose zeros in matrix entries which do not correspond to an edge.
}
