% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iModel-general.R
\name{logLik.iModel}
\alias{logLik.iModel}
\title{logLik.iModel <- function(object,...)
  structure(object$fitinfo$logL, df=object$fitinfo$dimension["df"], class="logLik")
extractAIC.iModel <- function(fit, scale, k = 2, ...){
  unname(c(fit$fitinfo$dimension["df"],
           fit$fitinfo$dev - k*fit$fitinfo$dimension["df"]))
}
AIC.iModel <- function(object, ..., k=2){
    extractAIC(object)[2]
}
## FIXME: need BIC for iModel
BIC.dModel <- function(object, ..., k=2){
    extractAIC(object, k=log(sum(object$datainfo$data)) )[2]
}}
\usage{
\method{logLik}{iModel}(object, ...)
}
\description{
logLik.iModel <- function(object,...)
  structure(object$fitinfo$logL, df=object$fitinfo$dimension["df"], class="logLik")
extractAIC.iModel <- function(fit, scale, k = 2, ...){
  unname(c(fit$fitinfo$dimension["df"],
           fit$fitinfo$dev - k*fit$fitinfo$dimension["df"]))
}
AIC.iModel <- function(object, ..., k=2){
    extractAIC(object)[2]
}
## FIXME: need BIC for iModel
BIC.dModel <- function(object, ..., k=2){
    extractAIC(object, k=log(sum(object$datainfo$data)) )[2]
}
}

