% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iModel-mmod.R
\name{mmod}
\alias{coef.mModel}
\alias{coefficients.mModel}
\alias{mmod}
\alias{mmod_dimension}
\alias{print.mModel}
\alias{summary.mModel}
\title{Mixed interaction model.}
\usage{
mmod(formula, data, marginal = NULL, fit = TRUE, details = 0)
}
\arguments{
\item{data}{}

\item{marginal}{}

\item{fit}{}

\item{details}{}
}
\value{
An object of class \code{mModel} and the more general class
    \code{iModel}.
}
\examples{

### FIXME: To be written

}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\seealso{
\code{\link{dmod}} \code{\link{cmod}}
}
\keyword{models}

