% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi-parm-conversion.R
\name{parm-conversion}
\alias{parm-conversion}
\alias{parm_pms2ghk}
\alias{parm_ghk2pms}
\alias{parm_pms2phk}
\alias{parm_phk2ghk}
\alias{parm_phk2pms}
\alias{parm_ghk2phk}
\alias{parm_CGstats2mmod}
\alias{parm_moment2pms}
\title{Conversion between different parametrizations of mixed models}
\usage{
parm_pms2ghk(parms)

parm_ghk2pms(parms)

parm_pms2phk(parms)

parm_phk2ghk(parms)

parm_phk2pms(parms)

parm_ghk2phk(parms)

parm_CGstats2mmod(parms, type = "ghk")

parm_moment2pms(SS)
}
\arguments{
\item{parms}{Parameters of a mixed interaction model}

\item{type}{Output parameter type; either "ghk" or "pms".}

\item{SS}{List of moment parameters.}
}
\value{
Parameters of a mixed interaction model.
}
\description{
Functions to convert between canonical parametrization
(g,h,K), moment parametrization (p,m,S) and mixed
parametrization (p,h,K).
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{utilities}
