% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gm_formula.R
\name{parse_gm_formula}
\alias{parse_gm_formula}
\title{Parse graphical model formula}
\usage{
parse_gm_formula(formula, varnames, marginal = NULL, interactions = NULL)
}
\arguments{
\item{formula}{A right hand sided formula or a list.}

\item{varnames}{Specification of the variables.}

\item{marginal}{Possible specification of marginal (a set of
variables); useful in connection with model specification
shortcuts.}

\item{interactions}{The maximum order of interactions allowed;
useful in connection with model specification shortcuts.}
}
\description{
Parse graphical model formula to internal representation
}
