% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_grips.R
\name{utilities_grips}
\alias{utilities_grips}
\alias{logLik.gips_fit_class}
\alias{AIC.gips_fit_class}
\alias{BIC.gips_fit_class}
\alias{sigma.gips_fit_class}
\alias{concentration}
\alias{concentration.gips_fit_class}
\alias{print.gips_fit_class}
\alias{summary.gips_fit_class}
\alias{glance.gips_fit_class}
\title{Utilities for gRips}
\usage{
\method{logLik}{gips_fit_class}(object, ...)

\method{AIC}{gips_fit_class}(object, ..., k = 2)

\method{BIC}{gips_fit_class}(object, ...)

\method{sigma}{gips_fit_class}(object, ...)

concentration(object, ...)

\method{concentration}{gips_fit_class}(object, ...)

\method{print}{gips_fit_class}(x, ...)

\method{summary}{gips_fit_class}(object, ...)

glance.gips_fit_class(x, ...)
}
\arguments{
\item{object}{Model object.}

\item{...}{Additional arguments; currently not used.}

\item{k}{Penalty parameter for calculating AIC; only k=2 gives genuine AIC.}

\item{x}{Object to be printed.}
}
\description{
Utilities for gRips
}
