% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRim-package.R
\docType{class}
\name{dModel-class}
\alias{cModel-class}
\alias{dModel-class}
\alias{mModel-class}
\title{Class \code{"dModel"}}
\description{
Setting formal classes for dModel, cModel and mModel objects
}
\section{Objects from the Class}{
 A virtual Class: No objects may be created
from it.
}
\examples{

showClass("dModel")

}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{classes}

